Add mocks module
